* (Playbook) Adding missing file for the eclipse project used in RTBareBones * (Playbook) Bugfix: Now respects the SetSoundEnabled flag * (Playbook) Audio manager:  now kills all sfx when suspending to background, important when you have looping sounds playing... * (Win) Bugfix: Audiere audio system now respects the SetSoundEnabled flag * (Win) Bugfix: IsIpad() no longer returns yes when emulating a playbook in windows * The optional C_SURFACE_UNLOAD_TEXTURES define for debugging is now explained in BaseApp.cpp in code comments * RGB supported added to SoftSurface (before, it was just 8bit paletted and RGBA) * Added Surface::InitFromSoftSurface(), easy way to get a GL surface from a software one * Added JPG image support to Proton!  RTSimpleApp modified to support it, (to convert your own app, it just needs jpg lib files added (skip if you're using Irrlicht, already added), JPGSurfaceLoader.cpp,  and RT_JPG_SUPPORT defined) * RTPack no longer incorrectly reports "(-forcealpha causing us to leave the unnecessary alpha channel in)" messages * RTPack.exe can now tell if alpha is really used or not, by scanning the alpha channel for info.  I had disabled that feature and forget to re-enable it after I added the -forcealpha parm a while ago... * RTPack:  Added new -ultra_compress <0 to 100> parm.  .rttex will internally use .jpg.  Will be ignores for images that use the alpha channel.  This can make a HUGE difference in filesize, important for Android. To use this, you must have R_JPG_SUPPORT defined, it will yell at you if you don't. * RTPack:  -forcealpha parm renamed -force_alpha to match the style of the others.  Don't worry, it still works with -forcealpha also * RTPack: .rttex can now wrap a (padded if required) jpg or other type of image file.  The idea is the game won't care what kind of data it gets and will just "work", so you can easily create multiple levels of quality for different build types * RTPack: Texture header now supports a flag to tell RTPack not to compress it, useful for things like jpegs that are already compressed * All RTSimpleApp project files for all platforms updated with jpg support, and their media_build.bat file has had the "ultra_compress" flag added (internally, this causes jpg to sometimes be used) * Fixed the OSX builds * Cleaned up RTSimpleApp a bit - the scrolling text in the About screen is handled better, there is no graphical overlay now, it uses the RenderScissorComponent to clip the scrolling text to its window, much cleaner than the old way. ( RenderScissorComponent.cpp was added to all project files) * Added UnityBuild_IrrlichtGUI.cpp and UnityBuild_IrrlichtMain.cpp, if you want irrlicht and don't feel like adding all the files manually, use these. (Good for Android builds, stops errors with too many files in one command line) * Added the Playbook (BBX) AudioManager .cpp * Added some helpful .bat files to the /bbx (playbook) directory of RTSimpleApp, dealing with running and final signing for the store, maybe helpful to use as templates * BUGFIX: RenderScissorComponent now works properly even when emulating a different screen size (stretching to fit a larger screen but the game things its 480X320 for instance) * g_landScapeNoNeckHurtMode and its special cases were removed. Use the "iPhone Landscape" profile if you don't want to twist your neck while emulating in windows. * Added SetForceOrientation() and GetForceOrientation(), needed internally for iOS emulation in Windows * The Win change orientation hot keys portrait, landscape, up, down now work in all video modes unless orientation is forced * The change orientation hot keys now require the _DEBUG flag to work, and Ctrl to be held down as well * Addded bool NeedToUseFakeScreenSize() * Added CL_Rectf GetScreenRect() helper function * BUGFIX: Fixed possible crash if you were holding down the mouse/finger when the app exited on some components * Simplied and bug fixed some things dealing with screen size stretching mixed with rotations

+++ shared/OSX/app/MyApplication.mm
@@ -25,12 +25,12 @@ bool g_controlDown = false;
 //	LogMsg("Sending key %d as %d", key, int(keyDown));
 	if ( keyDown)
 	{
-		GetMessageManager()->SendGUI(MESSAGE_TYPE_GUI_CHAR_RAW,   key, 1);  	
-		GetMessageManager()->SendGUI(MESSAGE_TYPE_GUI_CHAR,   key, 0);  	
+		GetMessageManager()->SendGUI(MESSAGE_TYPE_GUI_CHAR_RAW,   (float)key, 1.0f);  	
+		GetMessageManager()->SendGUI(MESSAGE_TYPE_GUI_CHAR,   (float)key, 0.0f);  	
 		
 	} else
 	{
-		GetMessageManager()->SendGUI(MESSAGE_TYPE_GUI_CHAR_RAW,   key, 0);  	
+		GetMessageManager()->SendGUI(MESSAGE_TYPE_GUI_CHAR_RAW,   (float)key, 0.0f);  	
 	}
 	
 }
